+++ packages/kit-headless/src/components/accordion/index.ts
@@ -1,4 +1,4 @@
-export { HAccordionRoot as Root } from './accordion-root';
+export { HAccordionRoot as Root } from './accordion-inline';
 export { HAccordionItem as Item } from './accordion-item';
 export { HAccordionTrigger as Trigger } from './accordion-trigger';
 export { HAccordionContent as Content } from './accordion-content';
